dts: arm64: intel_socfpga: Add peripherals register map
This patch is to add the peripheral register map in dts as syscon
to cover the peripheral for UART, I2C, I2C_EMAC, and Timer SP
addresses to be added in the MMU region. Due to the 4K alignment
checking in MMU driver, the peripheral that is not 4K aligned
will cause the kernel fall into kernel panic during MMU setup
when the CONFIG_ASSERT option is turned on.

Signed-off-by: Boon Khai Ng <boon.khai.ng@intel.com>